backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-environ
psycopg2-binary
django-cors-headers
pytest